feat: 🎸 times() accepts an object too
Sometimes you want to use `times` to create a random number of items.
You can now pass an object with the properties `min` and `max`. `times`
will return an array of objects of length between `min` and `max`. The
default values are `0` for `min` and `10` for `max`.

BREAKING CHANGE: 🧨 `times` will throw if you don't pass a number or an object